Add a new error when signing raw transactions specific to multisig
For the signing of raw transactions that pay to a P2SH multisignature
output, a new error has been added to inform the user that not enough
signatures havebeen produced in place of the non-informative txscript
error 'stack underflow'. Note that both behaviours deviate from bitcore,
which simply results the partially signed transaction without any
errors.