Detect silent network drops.
This change introduces additional network activity with the btcd
process to ensure that the network connection is not silently dropped.
Previously, if the connection was lost (e.g. wallet runs on a laptop
and connects to remote btcd, and the laptop is suspended/resumed) the
lost connection would not be detectable since all normal RPC activity
(excluding requests from btcwallet to btcd made by the user) is in the
direction of btcd to wallet in the form of websocket notifications.

Rather than simply pinging the server and waiting for the response, a
unique session ID has been added for each websocket client connection.
These session IDs are randomly generated by the btcd RPC server when a
new websocket connection is established, and are queried using the
session RPC.  If the session ID changes, it is assumed that the
connection silently reconnected, and the client is shutdown for this
case as well.