Fix bug in the address pool when restoring from seed
The address pool would fail to restore to the correct index when
restoring from a wallet seed. This was fixed by adding better tracking
of the address index in the wallet. A large portion of the code in
address pool was refactored for both clarity and concision. Some
code handling storing of the next to use address in the address manager
was also refactored to expand its use case and to make its role in
syncing the address pool clearer.