Add getstakeinfo RPC call handling to the wallet
The getstakeinfo command has been added to the wallet. It gives various
statistics about the wallet, such as the number of tickets it has in
mempool, the number of tickets it has that are mature or live, the
number of votes it has cast, and the total rewards earned by the wallet.
New functions that enable this to be calculated have been added to the
stake manager.