Increase the performance of getstakeinfo
getstakeinfo caused issues on the stake pool because of very slow
performance. The speed of the RPC call has been improved by reading
more local information instead of fetching it from the remote daemon.
New RPC calls for getrawmempool and existslivetickets are used, which
are substantially faster than the previous RPC calls.