Update for blockconnected notification API changes.
The blockconnected notification was modified to include relevant
transactions for the wallet from the block in an array, rather than
sending these transactions as individual notifications, each with the
same block metadata, before the blockconnected notifiation is received
(signaling that there were no more relevant transactions from that
block).

This has the potential to increase wallet database consistency by
making changes for an entire block at a time rather than inserting
transactions from a block before the wallet is actually marked in sync
with that block.  These fixes are not being made at this time, but the
newer notification API will make these fixes easier since all of the
relevant transactions are available at the same point in the code.