Use wtxmgr for input selection.
This adds a new wtxmgr API for resumable, incremental input selection
that is performed under a single database transaction, and hooks up
the API for transaction creation in the wallet package.  The new
wtxmgr API is compatible with the API required by the txauthor package
(txauthor.InputSource), but requires an (implicit) type conversion so
the wtxmgr package doesn't introduce an unnecessary dependency on the
txauthor package.

Stake and multisig transaction creation is unmodified by this change.