Fix a bug causes wallet lockup when making transactions
A bug existed that could cause the unusual locking up of the wallet
when requesting a transaction, such as through sendtoaddress. The
new function MakeInputSource opens a database transaction, but when
the address pool runs out of addresses it also attempts to open a
database transaction through waddrmgr.NextInternalAddresses. This
would cause a gridlock. To fix this, the address is pulled before
the database transaction is opened and then committed to later if
the transaction succeeds and includes the change address.