Isolate address pool to prevent excessive address creation
Automatic ticket purchases by the wallet could cause excessive usage
of the default account internal branch because the address for the
split transaction outputs was created using the exported new address
function for the wallet. The API for txToOutputs has been modified and
a new function that calls it, TxToOutputs, created. This new function
does all the locking required, so that the internal function txToOutputs
may be called by other functions in wallet that already have the
relevant locks held.