Add stake pool mode to the wallet (#192)
A stake pool mode for the wallet has been added that only inserts
tickets paying out to an extended public key passed to the server.
The extended public key and the index of the public key to scan to
are passed as the new command line or configuration argument
"enablestakepool". Required fees for the pool are passed in the
previously added "poolfees" argument.

The "poolfees" argument has been changed to a float64 representing
a percentage of the subsidy paid to the pool.

New handling for stake pool user information has been added to the
meta bucket of the stake pool, to make the stake pool UI faster,
more informative, and more efficient.

The new RPC call 'addticket' has been added for the manual entry of
tickets to the stake pool.