Fix fee calculation for revocations and rebroastcast on start up (#254)
The fee calculation for revocations was stuck at a very high fixed
rate. The fees for revocations are now calculated on a per kilobyte
basis. The wallet now also triggers winner and missed ticket
notifications on start up, so that revocations for tickets that
were missed while the wallet was offline will be triggered when
the wallet is started.

Fixes #51.