Update the wallet to begin allowing extended votebits setting
This fixes the TODO where wallet could not do the setting of extended
vote bits.  A new function allowing for the setting of multiple vote
bits for many tickets was also added.  A configuration setting has
been added to allow for the setting of the extended vote bits, and the
default uint32 "version" for the wallet client has been started at
2, awaiting a soft fork where the stake voter consensus version is
included in the proof of work header.

The wallet is also updated to the latest dependencies.  Fixes #363.