Fully update PoolTickets when using AddTicket rpc (#365)
Prior to this there was no updating of PoolTicket state when manually using AddTicket for formerly invalid tickets. Now they are properly removed from the invalid slice and a new PoolTicket is added to the valid slice. This along with some more upgrades will help stakepool operators ensure that all of their wallets are in sync with one another.