Fix a bug in the semver compatibility check. (#380)
While here, improve the error returned by RPCClient.Connect to include
both the required and advertised API versions.  These versions are
independent of the program and/or project's release versions, but it
should still be useful in determining whether dcrd must be downgraded
or upgraded for compatibility with this wallet software.

Add a 30 second sleep after any connection or compatibility errors
(where compatibility errors are not just incorrect API versions, but
mismatched currency networks as well).  This prevents an immediate
reconnect to the server.

Fixes #379.