RPCClient is responsible for attempting to reconnect to the server on its own. If the connection drops and it is able to reconnect, it provides a continuous stream of notifications from the server as if there was no interruption.
Wallet is no longer responsible for closing RPCClient. Wallet and
RPCClient are treated as separated resources that are both managed
by the walletMain.

rpcClientConnectLoop has been removed. The only reason this bizarre
function was needed is that the Wallet always shuts down the
RPCClient if it is unloaded. There is no logical reason this should
happen in the first place.