wallet: use fast scrypt options for regtest and simnet
In unit tests we can pass in the fast scrypt options directly.
This is not true for integration tests of external projects that
use btcwallet. But we can speed up these tests by using the fast
scrypt options if the regtest or simnet chain params are used as
they should never be set in production wallet creation.