build: revert update to walletdb v1.3.0
The v1.3.0 version of the walletdb module contained a bug fix within
boltdb to allow the use of go1.14, but it uncovered a GC issue. Rather
than reverting the commit itself that upgraded the dependency, we opt to
revert the change ourselves since we'll need additional changes to
ensure we're pinned to the intended walletdb version (v1.2.0).