chain/neutrino: fix nil pointer dereference
This commit fixes a potential nil pointer dereference in the neutrino
chain client. Currently the Rescan method blindly overwrites the rescan
member with nil _after_ reacquiring the lock. However, the field may
have been populated with a new rescan while the mutex wasn't held,
causing a panic later on where we assume the rescan is non-nil.

To remedy, we add a simple check that asserts we are nilling the same
rescan we were shutting down while the mutex was unlocked. If we find a
differing value, we will continue to this process until we arrive at
matching rescan pointers, at which point we know we can safely proceed
in creating a new rescan.