legacyrpc: add address type arg to getNewAddress and getRawChangeAddress
This handles the AddressType argument in the getNewAddress RPC handler.
It recognizes "legacy", "p2sh-segwit", and "bech32" to match Bitcoin
Core's RPC options.  These correspond to bip44, bip49 "plus", and bip84.

These are the waddrmgr.DefaultKeyScopes.

The validateaddress RPC is already able to recognize these addresses,
and it may be used to verify the addresses returned by getnewaddress.

This adds btcjson.ErrRPCWalletInvalidAddressType with a code (-5) and
message ("unknown address type") to match Bitcoin Core's v22.