chain: resolve data race in sequential start stop
Resolve the data race in the existing unit test by removing the
variable aliasing on s.rescanErr in notification handler
and by only writing ClientConnected notification after the notification
handler is running.

Write ClientConnected notification via a select statement as a
goroutine is unnecessary to send to the unbounded queue.