chain: add NotifyReceived unit test
Add TestNeutrinoClientNotifyReceived to verify that calls to
NotifyReceived call the Update method of the rescan object.  Verify
that there is no race condition on any state variables of the
NeutrinoClient.  Pass test with -race flag enabled.

Add a newRescanner private property of the NeutrinoClient to
dynamically inject the rescan constructor.  Cannot test without dyanmic
dependency injection as the neutrino client previously expected a
*neutrino.ChainService struct to create the rescan object.  Test
with a mockRescan object requires that this assumption be overridden
for unit tests.