chain+wallet: skip rescan if address batch is empty
We add a shortcut to the bitcoind rescan that allows us to finish
immediately if there are no addresses to re-scan.
We document that btcd already does this on the chain backend side and
that no such shortcut needs to be added to the btcwallet code.

Neutrino is a bit more tricky and isn't done in this commit. That likely
warrants its own pull request and update of tests.