chain: handle replacement and duplicate txes properly
Previously when adding a replacement tx to the local mempool, when the
txid's nested map becomes empty, we forgot to delete the key from the
root map. And when adding the same tx to the `txids` map(which happens
in ZMQ), we may treat the tx as a replacement tx. These are now fixed
and the related unit tests are added/updated to reflect the fix.