chain: detect and use gettxspendingprevout RPC instead of mempool
This change uses the gettxspendingprevout RPC if the bitcoind version
is greater or equal to version 24.0.0. In this case, the mempool is
no longer used for zmq clients. The original behavior must be kept
for rpc polling clients since they notify on new transactions and can
only learn about them from polling the bitcoind mempool and keeping
track of its own btcwallet mempool.