wallet+chain: add new APIs to chain interface
In this commit, we add new APIs to chain interface
used across `btcd`, `neutrino`, and `bitcoind` chain
backends. The new APIs are `GetPeerInfo`,
`GetBlockChainInfo`, and `RawRequest`.

The motivation behind adding those APIs is to easier the
process of unifying RPC connection creations across chain
backends as part of LND Daemon.